requirement.txt
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-WTF==1.0.1
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
